+++ Editor/SpriteLib/SpriteLibraryEditor/UI/CustomElements/GridView.cs
@@ -813,6 +813,7 @@ namespace UnityEditor.U2D.Animation.SpriteLibraryEditor
         VisualElement CreateDummyItemElement()
         {
             var item = new VisualElement();
+            item.pickingMode = PickingMode.Ignore;
             SetupItemElement(item);
             return item;
         }
@@ -1340,6 +1341,7 @@ namespace UnityEditor.U2D.Animation.SpriteLibraryEditor
             public RecycledRow(float height)
             {
                 AddToClassList(k_RowUssClassName);
+                pickingMode = PickingMode.Ignore;
                 style.height = height;
 
                 indices = new List<int>();